requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
click==8.0.1
cycler==0.11.0; python_version >= '3.6'
flask-cors==3.0.10
flask==2.0.1
fonttools==4.28.4; python_version >= '3.7'
gunicorn==20.1.0
imageio==2.9.0
itsdangerous==2.0.1
jinja2==3.0.1
kiwisolver==1.3.2; python_version >= '3.7'
markupsafe==2.0.1
matplotlib==3.5.1
numpy==1.21.2
opencv-contrib-python-headless==4.5.4.60
packaging==21.3; python_version >= '3.6'
pillow==8.3.1
pyparsing==3.0.6; python_version >= '3.6'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
setuptools==59.6.0; python_version >= '3.6'
six==1.16.0
werkzeug==2.0.1
